move argument count validation to happen later
This should result in invocations with valid argument counts still
being executed and only invocations with invalid argument counts fail.

Issue #3708
